+++ src/main/java/io/nats/client/impl/NatsConnection.java
@@ -202,6 +202,8 @@ class NatsConnection implements Connection {
             return;
         }
 
+        this.writer.setReconnectMode(true);
+
         while (!isConnected() && !isClosed() && !this.isClosing()) {
             Collection<String> serversToTry = buildReconnectList();
 
@@ -243,8 +245,8 @@ class NatsConnection implements Connection {
         }
 
         this.subscribers.forEach((sid, sub) -> {
-            if (!sub.isDraining()) {
-                sendSubscriptionMessage(sub.getSID(), sub.getSubject(), sub.getQueueName());
+            if (sub.getDispatcher() == null && !sub.isDraining()) {
+                sendSubscriptionMessage(sub.getSID(), sub.getSubject(), sub.getQueueName(), true);
             }
         });
 
@@ -259,6 +261,8 @@ class NatsConnection implements Connection {
         } catch (Exception exp) {
             this.processException(exp);
         }
+        
+        this.writer.setReconnectMode(false);
 
         processConnectionEvent(Events.RESUBSCRIBED);
     }
@@ -324,7 +328,7 @@ class NatsConnection implements Connection {
                     this.timer.schedule(new TimerTask() {
                         public void run() {
                             if (isConnected()) {
-                                sendPing();
+                                softPing(); // The timer always uses the standard queue
                             }
                         }
                     }, pingMillis, pingMillis);
@@ -700,7 +704,7 @@ class NatsConnection implements Connection {
             protocolBuilder.append(String.valueOf(after));
         }
         NatsMessage unsubMsg = new NatsMessage(protocolBuilder.toString());
-        queueOutgoing(unsubMsg);
+        queueInternalOutgoing(unsubMsg);
     }
 
     // Assumes the null/empty checks were handled elsewhere
@@ -718,11 +722,11 @@ class NatsConnection implements Connection {
         sub = new NatsSubscription(sid, subject, queueName, this, dispatcher);
         subscribers.put(sid, sub);
 
-        sendSubscriptionMessage(sid, subject, queueName);
+        sendSubscriptionMessage(sid, subject, queueName, false);
         return sub;
     }
 
-    void sendSubscriptionMessage(CharSequence sid, String subject, String queueName) {
+    void sendSubscriptionMessage(CharSequence sid, String subject, String queueName, boolean treatAsInternal) {
         if (!isConnected()) {
             return;// We will setup sub on reconnect or ignore
         }
@@ -740,7 +744,12 @@ class NatsConnection implements Connection {
         protocolBuilder.append(" ");
         protocolBuilder.append(sid);
         NatsMessage subMsg = new NatsMessage(protocolBuilder.toString());
-        queueOutgoing(subMsg);
+
+        if (treatAsInternal) {
+            queueInternalOutgoing(subMsg);
+        } else {
+            queueOutgoing(subMsg);
+        }
     }
 
     String createInbox() {
@@ -965,14 +974,22 @@ class NatsConnection implements Connection {
         String connectOptions = this.options.buildProtocolConnectOptionsString(serverURI, info.isAuthRequired());
         connectString.append(connectOptions);
         NatsMessage msg = new NatsMessage(connectString.toString());
-        queueOutgoing(msg);
+        queueInternalOutgoing(msg);
+    }
+    
+    CompletableFuture<Boolean> sendPing() {
+        return this.sendPing(true);
     }
 
+    CompletableFuture<Boolean> softPing() {
+        return this.sendPing(false);
+    }
+    
     // Send a ping request and push a pong future on the queue.
     // futures are completed in order, keep this one if a thread wants to wait
     // for a specific pong. Note, if no pong returns the wait will not return
     // without setting a timeout.
-    CompletableFuture<Boolean> sendPing() {
+    CompletableFuture<Boolean> sendPing(boolean treatAsInternal) {
         int max = this.options.getMaxPingsOut();
 
         if (!isConnectedOrConnecting()) {
@@ -989,14 +1006,20 @@ class NatsConnection implements Connection {
         CompletableFuture<Boolean> pongFuture = new CompletableFuture<>();
         NatsMessage msg = new NatsMessage(NatsConnection.OP_PING);
         pongQueue.add(pongFuture);
-        queueOutgoing(msg);
+
+        if (treatAsInternal) {
+            queueInternalOutgoing(msg);
+        } else {
+            queueOutgoing(msg);
+        }
+
         this.statistics.incrementPingCount();
         return pongFuture;
     }
 
     void sendPong() {
         NatsMessage msg = new NatsMessage(NatsConnection.OP_PONG);
-        queueOutgoing(msg);
+        queueInternalOutgoing(msg);
     }
 
     // Called by the reader
@@ -1084,6 +1107,13 @@ class NatsConnection implements Connection {
             throw new IllegalArgumentException("Control line is too long");
         }
         this.writer.queue(msg);
+    }
+
+    void queueInternalOutgoing(NatsMessage msg) {
+        if (msg.getControlLineLength() > this.options.getMaxControlLine()) {
+            throw new IllegalArgumentException("Control line is too long");
+        }
+        this.writer.queueInternalMessage(msg);
     }
 
     void deliverMessage(NatsMessage msg) {